+++ public/index.html
@@ -25,7 +25,7 @@
   </header>
   <main class="news-container">
     <div class="container">
-      <h2>THE LTEST NEWS </h2>
+      <h2>THE LATEST NEWS </h2>
       <section class="main-content">
         <div class="news">
           <h3 class="news-h3">First topic</h3>
