--- public/script/logic.js
@@ -1,10 +1,0 @@
-const fetching = (data) => {
-  return fetch('/search', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(data),
-  }).then((res) => res.json())
-    .then((res) => console.log(res));
-};
