+++ src/app.js
@@ -1,9 +1,10 @@
 const express = require('express');
 const path = require('path');
-require('dotenv').config();
+const err = require('./handlers/error');
+const search = require('./handlers/search');
 
 const app = express();
-const fetch = require('node-fetch');
+
 app.use(express.json());
 app.set('port', process.env.PORT || 4000);
 
@@ -15,23 +16,9 @@ app.get('/search', (req, res) => {
   res.sendFile(path.join(__dirname, '..', 'public', 'index.html'));
 });
 
-app.post('/search', (req, res) => {
-  const data = req.body.value;
-  const url = `http://newsapi.org/v2/everything?q=${data}&apiKey=${process.env.apiKey}`;
-  fetch(url)
-    .then(res => res.json())
-    .then(body => res.json(body.articles));
-});
+app.post('/search', search);
 
-app.use((err, req, res, next) => {
-  res
-    .status(500)
-    .sendFile(path.join(__dirname, '..', 'public', 'error500.html'));
-});
+app.use(err.client);
+app.use(err.server);
 
-app.use((req, res) => {
-  res
-    .status(404)
-    .sendFile(path.join(__dirname, '..', 'public', 'error404.html'));
-});
 module.exports = app;
